web/requirements.txt
fastapi
uvicorn
kafka-python
sonyflake-py
cassandra-driver
pymemcache